services/langextract/requirements.txt
# LangExtract service requirements
langextract>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# AI Provider SDKs
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Utility libraries
aiofiles>=23.2.1
orjson>=3.9.0